requirements.txt
absl-py==0.13.0
accelerate==0.4.0
alignment==1.0.10
apex @ file:///home/yandex/AMNLP2021/eladb3/s2e-coref/apex
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bcj-cffi==0.5.1
biopython==1.79
bleach==4.0.0
Brotli==1.0.9
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
click==8.0.1
cycler==0.10.0
datasets==1.11.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.3
filelock==3.0.12
fsspec==2021.7.0
fuzzysearch==0.7.3
fuzzywuzzy==0.18.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.33.0
google-auth-oauthlib==0.4.4
grpcio==1.38.1
huggingface-hub==0.0.12
idna==3.2
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonlines==2.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lab==6.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
multiprocess==0.70.12.2
multivolumefile==0.2.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.3
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626271506491/work
oauthlib==3.1.1
olefile==0.46
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1625655817137/work
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
py7zr==0.16.1
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.10.1
Pygments==2.9.0
pyparsing==2.4.7
pyppmd==0.15.2
pyrsistent==0.18.0
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
pyzstd==0.14.4
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
rouge-score==0.0.4
rsa==4.7.2
sacremoses==0.0.45
scipy==1.5.2
Send2Trash==1.8.0
sentencepiece==0.1.96
simplejson==3.17.3
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==4.0.0
tensorboard==2.3.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.11.0
testpath==0.5.0
texttable==1.6.4
tokenizers==0.10.3
torch==1.9.0
torchaudio==0.9.0a0+33b2469
torchvision==0.10.0
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
transformers==4.9.2
txt2tags==3.7
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
xxhash==2.0.2
